ml/requirements.txt
openai-whisper==20231117
torch>=2.1.0
torchaudio>=2.1.0
tiktoken==0.5.2
numpy>=1.26.2
more-itertools>=10.1.0
tqdm>=4.66.1
ffmpeg-python>=0.2.0
pandas>=2.3.0